+++ apt/cmdline/apt-get.cc
@@ -578,7 +578,7 @@ bool TryToInstall(pkgCache::PkgIterator Pkg,pkgDepCache &Cache,
    if (Pkg->VersionList == 0 && Pkg->ProvidesList != 0)
    {
       vector<pkgCache::Package *> GoodSolutions;
-      unsigned long long Size = 0;
+      unsigned long Size = 0;
       for (pkgCache::PrvIterator Prv = Pkg.ProvidesList();
 	   Prv.end() == false; Prv++)
 	 Size++;
@@ -1050,7 +1050,7 @@ pkgSrcRecords::Parser *FindSrc(const char *Name,pkgRecords &Recs,
    
    // The best hit
    pkgSrcRecords::Parser *Last = 0;
-   unsigned long long Offset = 0;
+   unsigned long Offset = 0;
    string Version;
    bool IsMatch = false;
